Added ceph_monconfig_autodiscover type and provider.
This type use SRV record to query a service rather than A record.
I would make the same change in function also. The advantage of using SRV record is that we can have hash of both hostname as well as IP addresses rather than just an array of IP addresses