Redirect the maybe-upgrade.sh output to a logger
I am seeing a strange behavior, where when I am running maybe-upgrade.sh
manually, all validation checks are running file and updating its status in etcd
as successful validation, but on and off it is getting changed.

During these kind of scenario, it would be easier to get the result of this
script in log.