Create single disk raid0 on hp storage boxes
This is required on hp storage node as hp nodes will not let the operating
system see the disks without creating any RAID on those disks. This is the
workaround hp people suggested to do so that operating system detect the disks

This works like mentioned below:

1. Facter hp_unassigned_disks to detect all unassigned disks
2. rjil::ceph::osd to call hp_drive type with the above fact as parameter
3. hp_drive will accept drive id as parameter and make raid 0 on individual
disks and let them visible to operating system.
4. install hpacucli package

NOTE: currently hpacucli is not there in any of the repo we use, this package
need to be added before we can merge this code.