set default ceph pool page size to 128
currently, the default page size is 1024,
this means that this value is used for all
of the on-demand created pages. From a ceph
best practices perspective, this puts way more
pages per osd than recommended (we place about
7k instead of the recommended 300). While
this isn't the root cause of the performance
degradation that we've seen on ceph setup,
it will greatly reduce the time added by these
issues, buying us some more time to root-cause.