Create users, tenants and default networks
This patch is to create tenants, users, and optional default networks.

The motive for this patch is to provide a way to use
rjil::neutron::contrail::fip_pool to create a restricted fip pool and assign it
to specific tenants, so those tenants should be created from puppet. The default
network/subnet created may be useful, another reason to create the default
network because we may hit contrail bug 1353325.