Use runit for running services inside container
Upstart will not work with docker containers. Also docker doesnt run "init"
script, but run specific script (in our case /entrypoint.sh) as init. So
certain tasks like reaping child processes will not be done by this script.

As a solution for above two problems and to easily add services to container,
this patch add runit which is a service management system (which is more like
supervisord). Everything including the script which run puppet and validation
would be run as services in container.

Runit is a lightwait service management toolset which is lighter than upstart
(http://smarden.org/runit/).

runit will log the output with timestamp, so removed printing timestamp from
maintain.sh