Added support for common volumes for docker containers
Added option to have an array of volumes which will be mounted on all containers.
This is helpful to share any volume between docker host and the containers,
especially helpful in dev environment where one just need to change the data
on docker host which will be available to all containers in that host. Combining
with vagrant, it is going to be very helpful.

For now used this to share puppet folders so that puppet changes can be tested
without rebuilding the containers (just need to delete and rerun them).