at least one container should be up before starting registrator
If registrator is started before any of the container is up, registrator will
fail to start (may be it is desired or may be due to a bug which need tobe
checked), and the containers on that node fail to register any consul services
because registrator is down.
 This patch make sure at least one container will be started before
registrator start.