Initialize ceph osd disk separately
It is dangerous to run disk management like partitioning, formatting the
disk every time, so this path is to separate the osd disk initialization to
separate process. This is called on initializing the osd servers from userdata
and vagrant. And further disk management to be done as separate operational
tool.