Handle more complicated usecases for container provisioning
Sometimes, you would need to handle more complex usecases for provisioning the
containers. For example, for ceph osds, puppet code need to determine the osd
disks and create one container per disk with appropriate parameters.

So use a container_wrapper definied type which would be called by rjil::docker
with appropriate parameters -

helper_class - a helper class which would be used to provision the containers
rather than standard rjil::docker::container. This helper class may have more
code to handle specialzed situations.
For example, in case of ceph osds, puppet code have to detect the disks
(automatically or with hiera config) and run individual containers for each
disk with appropriate container parameters.

helper_type - same as helper_class, the only difference is instead of a class,
this would be a type (defined type or puppet native type).

This implementation is to help using standard interface to define containers in
different situations, and to reduce duplicate code.

This patch also changed the hiera data for existing containers to work with the
change.

It also added some extra parameters for rjil::docker::container for more
flexibility.