Setup cinder containers
This patch is to setup cinder in container.

But cinder services are erroring out with import error on eventlet.zmq, it seems
there are some issues with cinder package or something. Because eventlet.zmq is
common service but other services like keystone, glance are working fine.

cinder is not working now.