Working cinder!!
Using cinder user to run cinder in container causing cinder services not getting
zmq messages (should be because of some os level permission issues which need to
be checked later). So running cinder as root as of now when in containers - this
should not cause much security problem as container itself is un-previleged, so
the root user in the container will not have root access to base os or on any
other containers running on it.

Also including some openstack_zeromq classes to make sure the zmq libraries.